Update OpenAL soft and use callbacks for movie audio

+++ mve/mve_audio.h
@@ -7,7 +7,7 @@ void mveaudio_uncompress(short *buffer, unsigned char *data, int length);
 #define MVESND_U8 2
 
 void mvesnd_init_audio(int format, int samplerate, int stereo);
-void mvesnd_queue_audio_buffer(int len, short* data);
+void mvesnd_queue_audio_buffer(int len, uint8_t* data);
 void mvesnd_close();
 
 void mvesnd_pause();
